Added parsing of time intervals
The new function search_dates_interval returns a list of tuples with the substring representing the date or time interval and a dictionary with the parsed date, the beginning and end date in case it is a time interval and an is_range boolean indicating if it is a time interval.
If a parsed substring has the period 'week', 'month' or 'year' it is considered a time interval.