WIP add find_date function: finds the first date
similar to search_dates, but gets only the first date,
and more suitable to shorts strings. It uses a brute-force approach,
but has more predictable performance
(at least on shorter strings - we limit the length to 100),
and better quality in our tests, although still not perfect.

Not sure if it should be exposed in current form, but may be useful
for future development.